(PA-5794) Update Checkout GitHub Action
Prior to this commit, the GitHub Actions were using the long end-of-life
Checkout action.

This commit updates Checkout to the latest version possible. The version
used in install_puppet.yaml is limited to v3 (which is end-of-life)
because it uses a version of NodeJS that relies on a compiler version
newer than CentOS 7 or Ubuntu 18.04 can support.

Once we drop support for those platforms, we should be able to use
Checkout v4 everywhere.